+++ crates/scheduler/src/protocol.rs
@@ -148,6 +148,15 @@ fn process_event_<B: Board>(
         }
         #[cfg(not(feature = "board-api-storage"))]
         Api::PlatformClearStore(_) => return Err(Error::world(Code::NotImplemented)),
+        #[cfg(feature = "native")]
+        Api::AppletReboot(_) => return Err(Error::world(Code::NotImplemented)),
+        #[cfg(any(feature = "pulley", feature = "wasm"))]
+        Api::AppletReboot(applet_id) => {
+            let service::applet::AppletId = applet_id;
+            <service::AppletInstall as TransferKind<B>>::start(scheduler);
+            <service::AppletInstall as TransferKind<B>>::finish(scheduler);
+            reply::<B, service::AppletReboot>(());
+        }
         #[cfg(not(feature = "_test"))]
         _ => return Err(Error::internal(Code::NotImplemented)),
     }